rename: Layout.jsx, Header.jsx, Footer.jsx의 디렉토리 변경
