+++ src/client/src/components/common/layout/Footer.jsx
